pyproject.toml
[tool.poetry]
name = "strava-peaks"
version = "0.1.0"
description = ""
authors = ["Adam Saegebarth <adam.saegebarth@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
stravaio = "^0.0.9"
requests = "^2.23.0"
pyyaml = "^5.3.1"
flask = "^1.1.2"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
